rename password component for clarity
