+++ packages/ui/src/components/tailwind/PasswordInput/index.stories.tsx
